add maven central settings and documentation
