+++ src/test/resources/com/homeofthewizard/maven/plugins/vault/certificate.pem
